move templating and import cleanup
